// src/enums/category.ts
export enum Category {
  BadgeInfo = 'BadgeInfo',
  CoopSkinInfo = 'CoopSkinInfo',
  GearInfoClothes = 'GearInfoClothes',
  GearInfoHead = 'GearInfoHead',
  GearInfoShoes = 'GearInfoShoes',
  NamePlateBgInfo = 'NamePlateBgInfo',
  WeaponInfoMain = 'WeaponInfoMain',
  WeaponInfoSpecial = 'WeaponInfoSpecial',
}